* Refactored Graph into more concise namespaces * Reformatted solution
