+++ Common.DataStructures/Graphs/Actions/EdgeAction.cs
@@ -1,0 +1,7 @@
+﻿namespace Common.DataStructures.Graphs.Actions;
+
+using Common.DataStructures.Graphs.Contracts;
+
+public delegate void EdgeAction<TKey, TEdge>(
+    TEdge edge)
+    where TEdge : IEdge<TKey>;
