+++ Common.DataStructures/Graphs/Actions/NodeAction.cs
@@ -1,0 +1,4 @@
+﻿namespace Common.DataStructures.Graphs.Actions;
+
+public delegate void NodeAction<in TKey>(
+    TKey nodeKey);
