+++ Common.DataStructures/Graphs/Contracts/IEdgeListAndIncidenceGraph.cs
@@ -1,9 +1,8 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+public interface IEdgeListAndIncidenceGraph<TKey, TEdge>
+    : IEdgeListGraph<TKey, TEdge>,
+        IIncidenceGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    public interface IEdgeListAndIncidenceGraph<TKey, TEdge>
-        : IEdgeListGraph<TKey, TEdge>,
-            IIncidenceGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-    }
 }
