+++ Common.DataStructures/Graphs/Contracts/IEdgeListGraph.cs
@@ -1,11 +1,9 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
-{
-    public interface IEdgeListGraph<TKey, TEdge> 
-        : IEdgeSet<TKey, TEdge>, 
-            IGraph<TKey, TEdge>, 
-            INodeSet<TKey>
-        where TEdge : IEdge<TKey>
-    {
+﻿namespace Common.DataStructures.Graphs.Contracts;
 
-    }
+public interface IEdgeListGraph<TKey, TEdge>
+    : IEdgeSet<TKey, TEdge>,
+        IGraph<TKey, TEdge>,
+        INodeSet<TKey>
+    where TEdge : IEdge<TKey>
+{
 }
