+++ Common.DataStructures/Graphs/Contracts/IGraph.cs
@@ -1,11 +1,10 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+public interface IGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    public interface IGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
+    public bool IsDirected
     {
-        public bool IsDirected
-        {
-            get;
-        }
+        get;
     }
 }
