+++ Common.DataStructures/Graphs/Contracts/IIncidenceGraph.cs
@@ -1,23 +1,22 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+using System.Collections.Generic;
+
+public interface IIncidenceGraph<TKey, TEdge>
+    : IImplicitGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    using System.Collections.Generic;
+    bool ContainsEdge(
+        TKey sourceNodeKey,
+        TKey targetNodeKey);
 
-    public interface IIncidenceGraph<TKey, TEdge>
-        : IImplicitGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-        bool ContainsEdge(
-            TKey sourceNodeKey,
-            TKey targetNodeKey);
+    bool TryGetEdge(
+        TKey sourceNodeKey,
+        TKey targetNodeKey,
+        out TEdge? edge);
 
-        bool TryGetEdge(
-            TKey sourceNodeKey,
-            TKey targetNodeKey,
-            out TEdge? edge);
-
-        bool TryGetEdges(
-            TKey sourceNodeKey,
-            TKey targetNodeKey,
-            out IEnumerable<TEdge>? edges);
-    }
+    bool TryGetEdges(
+        TKey sourceNodeKey,
+        TKey targetNodeKey,
+        out IEnumerable<TEdge>? edges);
 }
