+++ Common.DataStructures/Graphs/Contracts/IMutableGraph.cs
@@ -1,15 +1,8 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+public interface IMutableGraph<TKey, TEdge>
+    : IGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    /// <summary>
-    /// Defines a <see cref="IGraph&lt;TVertex, TEdge&gt;"/> that can be cleared.
-    /// </summary>
-    public interface IMutableGraph<TKey, TEdge>
-        : IGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-        /// <summary>
-        /// Removes all edges and nodes.
-        /// </summary>
-        void Clear();
-    }
+    void Clear();
 }
