+++ Common.DataStructures/Graphs/Contracts/IMutableIncidenceGraph.cs
@@ -1,17 +1,18 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+using Common.DataStructures.Graphs.Predicates;
+
+public interface IMutableIncidenceGraph<TKey, TEdge>
+    : IMutableGraph<TKey, TEdge>,
+        IIncidenceGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    public interface IMutableIncidenceGraph<TKey, TEdge>
-        : IMutableGraph<TKey, TEdge>,
-            IIncidenceGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-        void ClearOutEdges(
-            TKey sourceNodeKey);
+    void ClearOutEdges(
+        TKey sourceNodeKey);
 
-        int RemoveOutEdge(
-            TKey sourceNodeKey,
-            EdgePredicate<TKey, TEdge> predicate);
+    int RemoveOutEdge(
+        TKey sourceNodeKey,
+        EdgePredicate<TKey, TEdge> predicate);
 
-        void TrimEdgeExcessMemory();
-    }
+    void TrimEdgeExcessMemory();
 }
