+++ Common.DataStructures/Graphs/Contracts/IMutableNodeAndEdgeListGraph.cs
@@ -1,11 +1,10 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+public interface IMutableNodeAndEdgeListGraph<TKey, TEdge>
+    : IMutableNodeListGraph<TKey, TEdge>,
+        IMutableEdgeListGraph<TKey, TEdge>,
+        IMutableNodeAndEdgeSet<TKey, TEdge>,
+        INodeAndEdgeListGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    public interface IMutableNodeAndEdgeListGraph<TKey, TEdge>
-        : IMutableNodeListGraph<TKey, TEdge>,
-            IMutableEdgeListGraph<TKey, TEdge>,
-            IMutableNodeAndEdgeSet<TKey, TEdge>,
-            INodeAndEdgeListGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-    }
 }
