+++ Common.DataStructures/Graphs/Contracts/IMutableNodeListGraph.cs
@@ -1,10 +1,9 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+public interface IMutableNodeListGraph<TKey, TEdge>
+    : IEdgeListGraph<TKey, TEdge>
+    , IMutableNodeSet<TKey>
+    , IMutableEdgeListGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
 {
-    public interface IMutableNodeListGraph<TKey, TEdge>
-        : IEdgeListGraph<TKey, TEdge>
-        , IMutableNodeSet<TKey>
-        , IMutableEdgeListGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
-    }
 }
