+++ Common.DataStructures/Graphs/Contracts/IMutableNodeSet.cs
@@ -1,26 +1,27 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
+﻿namespace Common.DataStructures.Graphs.Contracts;
+
+using Common.DataStructures.Graphs.Actions;
+using Common.DataStructures.Graphs.Predicates;
+using System.Collections.Generic;
+
+public interface IMutableNodeSet<TKey> : INodeSet<TKey>
 {
-    using System.Collections.Generic;
+    bool AddNode(
+        TKey nodeKey);
 
-    public interface IMutableNodeSet<TKey> : INodeSet<TKey>
-    {
-        bool AddNode(
-            TKey nodeKey);
+    int AddNodes(
+        IEnumerable<TKey> nodeKeys);
 
-        int AddNodes(
-            IEnumerable<TKey> nodeKeys);
+    public event NodeAction<TKey> NodeAdded;
 
-        public event NodeAction<TKey> NodeAdded;
+    public event NodeAction<TKey> NodeRemoved;
 
-        public event NodeAction<TKey> NodeRemoved;
+    bool RemoveNode(
+        TKey nodeKey);
 
-        bool RemoveNode(
-            TKey nodeKey);
+    int RemoveNodes(
+        IEnumerable<TKey> nodeKeys);
 
-        int RemoveNodes(
-            IEnumerable<TKey> nodeKeys);
-
-        int RemoveNodes(
-            NodePredicate<TKey> predicate);
-    }
+    int RemoveNodes(
+        NodePredicate<TKey> predicate);
 }
