+++ Common.DataStructures/Graphs/Contracts/INodeAndEdgeListGraph.cs
@@ -1,10 +1,8 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
-{
-    public interface INodeAndEdgeListGraph<TKey, TEdge>
-        : IEdgeListGraph<TKey, TEdge>,
-            INodeListGraph<TKey, TEdge>
-        where TEdge : IEdge<TKey>
-    {
+﻿namespace Common.DataStructures.Graphs.Contracts;
 
-    }
+public interface INodeAndEdgeListGraph<TKey, TEdge>
+    : IEdgeListGraph<TKey, TEdge>,
+        INodeListGraph<TKey, TEdge>
+    where TEdge : IEdge<TKey>
+{
 }
