+++ Common.DataStructures/Graphs/Contracts/INodeEdgeDictionary.cs
@@ -1,13 +1,12 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
-{
-    using System.Runtime.Serialization;
+﻿namespace Common.DataStructures.Graphs.Contracts;
 
-    public interface INodeEdgeDictionary<TKey, TEdge>
-        : ICloneable,
-            IDictionary<TKey, IEdgeList<TKey, TEdge>>,
-            ISerializable
-        where TEdge : IEdge<TKey>
-    {
-        public new INodeEdgeDictionary<TKey, TEdge> Clone();
-    }
+using System.Runtime.Serialization;
+
+public interface INodeEdgeDictionary<TKey, TEdge>
+    : ICloneable,
+        IDictionary<TKey, IEdgeList<TKey, TEdge>>,
+        ISerializable
+    where TEdge : IEdge<TKey>
+{
+    public new INodeEdgeDictionary<TKey, TEdge> Clone();
 }
