+++ Common.DataStructures/Graphs/Contracts/INodeListGraph.cs
@@ -1,10 +1,8 @@
-﻿namespace Common.DataStructures.Graphs.Contracts
-{
-    public interface INodeListGraph<TKey, TEdge>
-        : IIncidenceGraph<TKey, TEdge>, 
-            INodeSet<TKey>
-        where TEdge : IEdge<TKey>
-    {
+﻿namespace Common.DataStructures.Graphs.Contracts;
 
-    }
+public interface INodeListGraph<TKey, TEdge>
+    : IIncidenceGraph<TKey, TEdge>,
+        INodeSet<TKey>
+    where TEdge : IEdge<TKey>
+{
 }
