--- Common.DataStructures/Graphs/EdgePredicate.cs
@@ -1,8 +1,0 @@
-﻿namespace Common.DataStructures.Graphs
-{
-    using Common.DataStructures.Graphs.Contracts;
-
-    public delegate bool EdgePredicate<TKey, TEdge>(
-        TEdge edge)
-        where TEdge : IEdge<TKey>;
-}
