+++ Common.DataStructures/Graphs/Handlers/NodeEventHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace Common.DataStructures.Graphs.Handlers;
+
+using Common.DataStructures.Graphs.Events;
+
+public delegate void NodeEventHandler<TKey>(
+    object sender,
+    NodeEventArgs<TKey> eventArgs);
