--- Common.DataStructures/Graphs/NodeAction.cs
@@ -1,5 +1,0 @@
-﻿namespace Common.DataStructures.Graphs
-{
-    public delegate void NodeAction<in TKey>(
-        TKey nodeKey);
-}
