--- Common.DataStructures/Graphs/NodeEventHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace Common.DataStructures.Graphs
-{
-    public delegate void NodeEventHandler<TKey>(
-        object sender,
-        NodeEventArgs<TKey> eventArgs);
-}
