--- Common.DataStructures/Graphs/NodePredicate.cs
@@ -1,4 +1,0 @@
-﻿namespace Common.DataStructures.Graphs;
-
-public delegate bool NodePredicate<TKey>(
-    TKey nodeKey);
