+++ Common.DataStructures/Graphs/Predicates/EdgePredicate.cs
@@ -1,0 +1,7 @@
+﻿namespace Common.DataStructures.Graphs.Predicates;
+
+using Common.DataStructures.Graphs.Contracts;
+
+public delegate bool EdgePredicate<TKey, TEdge>(
+    TEdge edge)
+    where TEdge : IEdge<TKey>;
