+++ Common.DataStructures/Graphs/Predicates/NodePredicate.cs
@@ -1,0 +1,4 @@
+﻿namespace Common.DataStructures.Graphs.Predicates;
+
+public delegate bool NodePredicate<TKey>(
+    TKey nodeKey);
