+++ Common.ExtensionLibrary.Tests/StringExtensionsTests.cs
@@ -1,105 +1,104 @@
-﻿namespace Common.ExtensionLibrary.Tests
+﻿namespace Common.ExtensionLibrary.Tests;
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Shouldly;
+
+[TestClass]
+public class StringExtensionsTests
 {
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-    using Shouldly;
-
-    [TestClass]
-    public class StringExtensionsTests
+    [TestMethod]
+    public void IsEqualTo_CaseSensitive_ReturnsFalse()
     {
-        [TestMethod]
-        public void IsEqualTo_CaseSensitive_ReturnsFalse()
-        {
-            string source = " TeSt";
-            string target = "tEsT ";
+        string source = " TeSt";
+        string target = "tEsT ";
 
-            var result = source.IsEqualTo(
-                target,
-                isCaseSensitive: true);
+        bool result = source.IsEqualTo(
+            target,
+            isCaseSensitive: true);
 
-            result.ShouldBeFalse();
-        }
+        result.ShouldBeFalse();
+    }
 
-        [TestMethod]
-        public void IsEqualTo_DefaultComparison_IgnoresCaseAndWhiteSpace()
-        {
-            string source = " TeSt";
-            string target = "tEsT ";
+    [TestMethod]
+    public void IsEqualTo_DefaultComparison_IgnoresCaseAndWhiteSpace()
+    {
+        string source = " TeSt";
+        string target = "tEsT ";
 
-            var result = source.IsEqualTo(target);
+        bool result = source.IsEqualTo(target);
 
-            result.ShouldBeTrue();
-        }
+        result.ShouldBeTrue();
+    }
 
-        [TestMethod]
-        public void IsEqualTo_NullValues_ReturnTrue()
-        {
-            string? source = null;
-            string? target = null;
+    [TestMethod]
+    public void IsEqualTo_NullValues_ReturnTrue()
+    {
+        string? source = null;
+        string? target = null;
 
-            var result = source.IsEqualTo(target);
+        bool result = source.IsEqualTo(target);
 
-            result.ShouldBeTrue();
-        }
+        result.ShouldBeTrue();
+    }
 
-        [TestMethod]
-        public void IsEqualTo_WhiteSpaceSensitive_ReturnsFalse()
-        {
-            string source = " TeSt";
-            string target = "tEsT ";
+    [TestMethod]
+    public void IsEqualTo_WhiteSpaceSensitive_ReturnsFalse()
+    {
+        string source = " TeSt";
+        string target = "tEsT ";
 
-            var result = source.IsEqualTo(
-                target,
-                isWhiteSpaceSensitive: true);
+        bool result = source.IsEqualTo(
+            target,
+            isWhiteSpaceSensitive: true);
 
-            result.ShouldBeFalse();
-        }
+        result.ShouldBeFalse();
+    }
 
-        [DataRow(null, true)]
-        [DataRow("", true)]
-        [DataRow(" ", false)]
-        [DataRow("test", false)]
-        [TestMethod]
-        public void IsNullOrEmpty(
-            string? value,
-            bool expectedResult)
-        {
-            var result = value.IsNullOrEmpty();
+    [DataRow(null, true)]
+    [DataRow("", true)]
+    [DataRow(" ", false)]
+    [DataRow("test", false)]
+    [TestMethod]
+    public void IsNullOrEmpty(
+        string? value,
+        bool expectedResult)
+    {
+        bool result = value.IsNullOrEmpty();
 
-            result.ShouldBe(expectedResult);
-        }
+        result.ShouldBe(expectedResult);
+    }
 
-        [DataRow(null, true)]
-        [DataRow("", true)]
-        [DataRow(" ", true)]
-        [DataRow("test", false)]
-        [TestMethod]
-        public void IsNullOrWhiteSpace(
-            string? value,
-            bool expectedResult)
-        {
-            var result = value.IsNullOrWhiteSpace();
+    [DataRow(null, true)]
+    [DataRow("", true)]
+    [DataRow(" ", true)]
+    [DataRow("test", false)]
+    [TestMethod]
+    public void IsNullOrWhiteSpace(
+        string? value,
+        bool expectedResult)
+    {
+        bool result = value.IsNullOrWhiteSpace();
 
-            result.ShouldBe(expectedResult);
-        }
+        result.ShouldBe(expectedResult);
+    }
 
-        [TestMethod]
-        public void SafeTrim_NullValue_ReturnsDefaultValue()
-        {
-            string? value = null;
+    [TestMethod]
+    public void SafeTrim_NullValue_ReturnsDefaultValue()
+    {
+        string? value = null;
 
-            var result = value.SafeTrim("default");
+        string? result = value.SafeTrim("default");
 
-            result.ShouldBe("default");
-        }
+        result.ShouldBe("default");
+    }
 
-        [TestMethod]
-        public void SafeTrim_NullValueWithoutDefault_ReturnsEmptyString()
-        {
-            string? value = null;
+    [TestMethod]
+    public void SafeTrim_NullValueWithoutDefault_ReturnsEmptyString()
+    {
+        string? value = null;
 
-            var result = value.SafeTrim();
+        string? result = value.SafeTrim();
 
-            result.ShouldBe(string.Empty);
-        }
+        result.ShouldBe(string.Empty);
     }
 }
